Actualización módulo de horarios, cursos, asignación docente

+++ SchoolSystem.Infrastructure/Respositories/HorarioRepository.cs
@@ -5,10 +5,9 @@ using SchoolSystem.Infrastructure.Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace SchoolSystem.Infrastructure.Respositories
+namespace SchoolSystem.Infrastructure.Repositories
 {
     public class HorarioRepository : IHorarioRepository
     {
@@ -18,19 +17,20 @@ namespace SchoolSystem.Infrastructure.Respositories
         {
             _context = context;
         }
+
         public async Task<bool> ExisteCruce(int docenteId, string dia, int horaLectiva, int periodoId)
         {
             return await _context.Horario
                 .AnyAsync(h => h.DiaSemana == dia &&
-                               h.HoraLectivaId == horaLectiva && 
-                               h.AsignacionDocente!.DocenteId == docenteId &&
-                               h.AsignacionDocente.PeriodoAcademicoId ==  periodoId);
+                               h.HoraLectivaId == horaLectiva &&
+                               h.AsignacionDocente!.DocenteId == docenteId && 
+                               h.AsignacionDocente.PeriodoAcademicoId == periodoId);
         }
 
         public async Task<bool> ExisteCruceSeccion(int seccionId, string dia, int horaLectivaId, int periodoId)
         {
             return await _context.Horario
-                .AnyAsync(h => h.DiaSemana == dia && 
+                .AnyAsync(h => h.DiaSemana == dia &&
                                h.HoraLectivaId == horaLectivaId &&
                                h.AsignacionDocente!.SeccionId == seccionId &&
                                h.AsignacionDocente.PeriodoAcademicoId == periodoId);
@@ -38,22 +38,38 @@ namespace SchoolSystem.Infrastructure.Respositories
 
         public async Task InsertarRangoAsync(IEnumerable<Horario> horarios)
         {
-            await _context.Horario.AddRangeAsync(horarios);
+            if (horarios != null && horarios.Any())
+            {
+                var horariosParaGuardar = horarios.Select(h => new Horario
+                {
+                    AsignacionDocenteId = h.AsignacionDocenteId,
+                    HoraLectivaId = h.HoraLectivaId,
+                    DiaSemana = h.DiaSemana
+                }).ToList();
+
+                await _context.Horario.AddRangeAsync(horariosParaGuardar);
+                await _context.SaveChangesAsync();
+            }
         }
 
         public async Task LimpiarHorariosPeriodoAsync(int periodoId)
         {
-            var horariosExiste = _context.Horario.Where(h => h.AsignacionDocente!.PeriodoAcademicoId == periodoId);
+            var horariosExiste = await _context.Horario
+                .Where(h => h.AsignacionDocente!.PeriodoAcademicoId == periodoId)
+                .ToListAsync();
 
-            _context.Horario.RemoveRange(horariosExiste);
-            await _context.SaveChangesAsync();
+            if (horariosExiste.Any())
+            {
+                _context.Horario.RemoveRange(horariosExiste);
+                await _context.SaveChangesAsync();
+            }
         }
 
         public async Task<List<Horario>> ObtenerPorGradoSeccionPeriodo(int gradoId, int seccionId, int periodoId)
         {
             return await _context.Horario
                 .Include(h => h.AsignacionDocente)
-                    .ThenInclude(a => a!.Curso)
+                    .ThenInclude(a => a!.PlanEstudio!.Curso)
                 .Include(h => h.AsignacionDocente)
                     .ThenInclude(a => a!.Docente)
                 .Include(h => h.AsignacionDocente)
@@ -63,6 +79,7 @@ namespace SchoolSystem.Infrastructure.Respositories
                 .Where(h => h.AsignacionDocente!.GradoId == gradoId &&
                             h.AsignacionDocente.SeccionId == seccionId &&
                             h.AsignacionDocente.PeriodoAcademicoId == periodoId)
+                .AsNoTracking()
                 .ToListAsync();
         }
     }
